いったんPartyManagerからDataManagerに移植
