+++ Assets/Script/DataBase/Test/FieldTest.cs
@@ -11,7 +11,7 @@ public class FieldTest : MonoBehaviour
         // スペースキーを押したらバトルシーンへ移動
         if (Input.GetKeyDown(KeyCode.Space))
         {
-            if (PartyManager.Instance != null)
+            if (DataManager.Instance != null)
             {
                 Debug.Log("サンプルシーンに移行");
                 FadeManager.FadeChangeScene("SampleScene", 1.0f);
@@ -20,9 +20,9 @@ public class FieldTest : MonoBehaviour
         // テスト用でダメージ与える
         if (Input.GetKeyDown(KeyCode.D))
         {
-            if(PartyManager.Instance != null && PartyManager.Instance.partyMembers.Count > 0)
+            if(DataManager.Instance != null && DataManager.Instance.partyMembers.Count > 0)
             {
-                Player leader = PartyManager.Instance.partyMembers[0];
+                Player leader = DataManager.Instance.partyMembers[0];
                 Debug.Log($"減少前HP {leader.GetStatusRuntime().hp}");
                 leader.TakeDamage(10);
                 Debug.Log($"減少後HP {leader.GetStatusRuntime().hp}");
@@ -33,7 +33,7 @@ public class FieldTest : MonoBehaviour
 
     void Start()
     {
-        if (PartyManager.Instance != null)
+        if (DataManager.Instance != null)
         {
 
         }
