--- Assets/Script/DataBase/Test/PartyManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4af24f08d88a3664b93729efd7f55ecb
